NULL repeated fields should be converted to an empty list
According to the 'JSON Mapping' section in the proto3 Language guide
found here:

https://developers.google.com/protocol-buffers/docs/proto3#json

It is mentioned that NULL repeated fields should be converted to an
empty list.

Currently protobuf2json-c does not convert a NULL repeated field to
JSON (it simply ignores it). This patch updates the library to ensure
that NULL repeated fields are converted to empty JSON lists.